Started working on star coloring.
I'm going to come back to this, I just wanted something more than plain
white. But the plan is to have the star's color determined by the size
of it, providing an effect similar to the main sequence. It is obviously
not going to be exact, but neither is the gravity simulated here sooooo.