Trying to get shiney points
My thought process is that the closer the collision to the center of the
star the brighter it ought to be. So I took the collision normal and
attempted to divide it by the ray thinking that would give me a
percentage of how far of it is. I don't think that is how it works. I'm
going to have to do some further considerations for next week to get
this idea off the ground. After I've got the gradient working correctly
I can then move on to transparency.