Following Realtime Rendering's guidance
Page 740 of my book on realtime rendering has some guidance on
raytracing. It is helping but I've clealy made a mistake some where.
While what I am getting seems to be in a general ballpark, it is
behaving as though all the rays are intersecting the same sphere or
something. The closer to the top left a star is the brighter it is, and
the closer a part of a star is tot he top left the brighter it is. Not
the behavior I was looking for.